+++ src/tests/cairo_run_test.rs
@@ -409,6 +409,13 @@ fn blake2s_hello_world_hash() {
 
 #[test]
 #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
+fn example_blake2s() {
+    let program_data = include_bytes!("../../cairo_programs/example_blake2s.json");
+    run_program_simple_with_memory_holes(program_data.as_slice(), 2);
+}
+
+#[test]
+#[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
 fn finalize_blake2s() {
     let program_data = include_bytes!("../../cairo_programs/finalize_blake2s.json");
     run_program_simple_with_memory_holes(program_data.as_slice(), 20);
@@ -720,7 +727,7 @@ fn uint384() {
 #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
 fn uint384_extension() {
     let program_data = include_bytes!("../../cairo_programs/uint384_extension_test.json");
-    run_program_simple_with_memory_holes(program_data.as_slice(), 20);
+    run_program_simple_with_memory_holes(program_data.as_slice(), 40);
 }
 
 #[test]
@@ -846,7 +853,7 @@ fn assert_le_felt_old() {
 #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
 fn cairo_run_fq_test() {
     let program_data = include_bytes!("../../cairo_programs/fq_test.json");
-    run_program_simple_with_memory_holes(program_data.as_slice(), 122);
+    run_program_simple_with_memory_holes(program_data.as_slice(), 120);
 }
 
 #[test]